Coding philosophy rules added
Section 2 of coding rules concerning code philosophy has been added to the document.
Foreword style change.
List of Tables and List of Figures are removed (I think I won't usethem in the future).
First page font harmonization with the rest of the document : was Helvetica, now Bollinium.
"Exemple"/"Note"/"Exception" for rules style change for small caps. Makes it easier to see in the text.